# cellp/api/openapi.yaml
openapi: 3.0.3
info:
  title: cellp API
  version: 1.0.0
  description: cellp control plane REST API (DESIGN §9)
servers:
  - url: http://127.0.0.1:8790/v1
paths:
...
  /projects/{projectID}/versions/{versionID}/workflows/{name}/instances:
    get:
      summary: List workflow instances via celld cell list
      description: |
        Read-only wrapper around `celld cell list [CLASS] --all --json` on the version bucket.
        `{name}` matches wrangler workflow `name`, then `binding`. Unknown name → 404 `workflow_not_found`
        (celld is not invoked). Missing celld → 200 with empty `instances`.
        When the version declares two or more workflows, `filter` is `script` and `limitation` explains
        that celld `cell list` 只给出 `__Workflow.{script}:<64-hex>`；实例内部名 `{workflow_name}/{instance_id}`
        被哈希，无法按 workflow 名精确过滤。本响应返回该 version / 该 script 下全部 Workflow cell，并附 wrangler
        声明的 workflow 名。精确归属需等 `celld workflow`。
        Pause / resume / restart / delete are not offered (no `celld workflow`).
      operationId: listWorkflowInstances
      security:
        - adminToken: []
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
        - name: versionID
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: wrangler workflows[].name, or binding if name does not match
      responses:
        "200":
          description: Instance list (possibly script-level fallback)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowInstances"
        "404":
          description: version_not_found | version_not_ready | bindings_not_found | workflow_not_found
        "502":
          description: celld_cell_list_failed
...
components:
  schemas:
...
    WorkflowInstances:
      type: object
      properties:
        workflow_name:
          type: string
        binding:
          type: string
        script_name:
          type: string
        filter:
          type: string
          enum: [workflow, script]
          description: workflow = attributable to this wrangler workflow; script = all Workflow cells for the script
        limitation:
          type: string
          nullable: true
          description: |
            Present when filter=script. celld `cell list` 只给出 `__Workflow.{script}:<64-hex>`；实例内部名
            `{workflow_name}/{instance_id}` 被哈希，无法按 workflow 名精确过滤。本响应返回该 version / 该 script
            下全部 Workflow cell，并附 wrangler 声明的 workflow 名。精确归属需等 `celld workflow`。
        wrangler_workflows:
          type: array
          items:
            type: string
        instances:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/WorkflowInstance"
    WorkflowInstance:
      type: object
      properties:
        scope:
          type: string
        class:
          type: string
        id:
          type: string
        reserved:
          type: boolean
  securitySchemes:
    deployToken:
      type: http
      scheme: bearer
      description: CELLP_DEPLOY_TOKEN or PLATFORM_TOKEN — POST /versions only
    adminToken:
      type: http
      scheme: bearer
      description: CELLP_ADMIN_TOKEN or PLATFORM_TOKEN (dev-local-token fallback) — full admin